be2/pyproject.toml
[project]
name = "be2"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "google-auth>=2.48.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "PyJWT>=2.8.0",
]

[tool.fastapi]
dev_port = 8001
